+++ Chapter8/src/exercise1/Staff.java
@@ -7,7 +7,7 @@ public class Staff extends UEmployee {
 	
 	private String jobTitle; 
 
-		public Staff(String Fn, String Ln, double Sal, String Job) {
+		public Staff(String Fn, String Ln, int Sal, String Job) {
 			super(Fn, Ln, Sal);
 			
 			jobTitle = Job;
